+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ## FlexMessage 0.0.2
 This new release introduces `CHANGELOG.md` :-)
 ### FEATURES
-  * Reset function for `FlexMessage` to reset all notifications
+- Reset function for `FlexMessage` to reset all notifications
 
 ### BUGFIXES
-  * Fix for `Gosec` Workflow
+- Fix for `Gosec` Workflow
 
 ## FlexMessage 0.0.1
-  * Initial release
+- Initial release